+++ CHANGELOG.md
@@ -14,6 +14,8 @@ All notable changes to this project will be documented in this file.
 -   View all profiles (index)
 -   View profile based on user id
 -   Delete profile & associated user
+-   Add a favorite restaurant to profile
+-   Delete favorite restaurant from profile
 
 ## Fixed